deactivated_user: Change the empty view text and compose_box.
If a user is deactivated, change empty PM/group PM view text
and keep the compose_box closed and blocked for the PM with
deactivated user.

Created a function inside `people.js` to verify,
if the given user_id is active or not, returns boolean.

Inside compose_actions.js, wrote a function that returns a boolean
indicating whether there is a deactivated user in a group or single
private message, as well as the number of users in the array of users
to be private messaged with.

To block the compose_box in case of deactivated users and expand it in
case of stream/private button, a function was written to check the
opts.trigger and msg_type.

The compose box needs to be disabled in all the cases where a deactivated user
is present in pm-with/ group-pm-with or all message narrow. To make this functional
update_reply_recipient_label is a function that is called everytime a selected_message
changes or user switches to pm. This is the same behaviour what disable the compose_box
is expected, to make this work, called a function inside update_reply_recipient_label to
update the status of the compose box and added few if-else conditions since on-reloading
certain properites are not present instantaneously to execute the function thus return
the normal behaviour.

Changed the `/frontend_tests/node_tests/compose_actions.js`,
according to the new changes made inside the `compose_actions.js`.
Changed the /node_tests/people.js, according to the function
created in people.js.
Changed the /node_tests/narrow.js, according to the changes made
in narrow_banner.js.

Fixes: #23408.