custom_profile_fields: Fix empty field choices issue of select field.
While editing select type profile field, if we submit zero choices
it fails silently, current implementation for this issue does not
work for some cases, and this commit will fix that.

It will disable the last choice and hide its delete button, so the
user can't clear or delete that choice.